Implement cli for AC launcher
